+++ src/service/search/tantivy/puffin/reader.rs
@@ -41,7 +41,7 @@ impl PuffinBytesReader {
     pub async fn read_blob_bytes(
         &self,
         blob_metadata: &BlobMetadata,
-        range: Option<core::ops::Range<usize>>,
+        range: Option<core::ops::Range<u64>>,
     ) -> Result<bytes::Bytes> {
         let raw_data = infra::cache::storage::get_range(
             &self.account,
@@ -82,12 +82,9 @@ impl PuffinBytesReader {
         }
 
         // check MAGIC
-        let magic = infra::cache::storage::get_range(
-            &self.account,
-            &self.source.location,
-            0..MAGIC_SIZE as usize,
-        )
-        .await?;
+        let magic =
+            infra::cache::storage::get_range(&self.account, &self.source.location, 0..MAGIC_SIZE)
+                .await?;
         ensure!(magic.to_vec() == MAGIC, anyhow!("Header MAGIC mismatch"));
 
         let puffin_meta = PuffinFooterBytesReader::new(self.account.clone(), self.source.clone())
@@ -123,7 +120,7 @@ impl PuffinFooterBytesReader {
 impl PuffinFooterBytesReader {
     async fn parse(mut self) -> Result<PuffinMeta> {
         // read footer
-        if self.source.size < FOOTER_SIZE as usize {
+        if self.source.size < FOOTER_SIZE {
             return Err(anyhow!(
                 "Unexpected footer size: expected size {} vs actual size {}",
                 FOOTER_SIZE,
@@ -133,7 +130,7 @@ impl PuffinFooterBytesReader {
         let footer = infra::cache::storage::get_range(
             &self.account,
             &self.source.location,
-            (self.source.size - FOOTER_SIZE as usize)..self.source.size,
+            (self.source.size - FOOTER_SIZE)..self.source.size,
         )
         .await?;
 
@@ -165,7 +162,7 @@ impl PuffinFooterBytesReader {
         self.payload_size = i32::from_le_bytes(payload_size) as u64;
 
         // read the payload
-        if self.source.size < FOOTER_SIZE as usize + self.payload_size as usize {
+        if self.source.size < FOOTER_SIZE + self.payload_size {
             return Err(anyhow!(
                 "Unexpected payload size: expected size {} vs actual size {}",
                 FOOTER_SIZE + self.payload_size,
@@ -175,10 +172,8 @@ impl PuffinFooterBytesReader {
         let payload = infra::cache::storage::get_range(
             &self.account,
             &self.source.location,
-            (self.source.size
-                - FOOTER_SIZE as usize
-                - self.payload_size as usize
-                - MAGIC_SIZE as usize)..(self.source.size - FOOTER_SIZE as usize),
+            (self.source.size - FOOTER_SIZE - self.payload_size - MAGIC_SIZE)
+                ..(self.source.size - FOOTER_SIZE),
         )
         .await?;
